update javadocs and code based on comments
