+++ src/main/java/ratspeak/exception/DukeException.java
@@ -1,7 +1,6 @@
-package duke.exception;
+package ratspeak.exception;
 
 public class DukeException extends Exception {
-
     public DukeException(String message) {
         super(message);
     }
