+++ src/main/java/ratspeak/gui/Launcher.java
@@ -1,4 +1,4 @@
-package duke.gui;
+package ratspeak.gui;
 import javafx.application.Application;
 
 /**
