+++ src/main/java/ratspeak/task/Deadline.java
@@ -1,6 +1,6 @@
-package duke.task;
+package ratspeak.task;
 
-import duke.exception.DukeException;
+import ratspeak.exception.DukeException;
 
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
